services/upload_service/requirements.txt
FLASK
psycopg2-binary
minio